requirements.txt
Django
psycopg2
markdown
arrow
djangorestframework
django-filter
django_chartit
django-crispy-forms
plotly
numpy
